% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickett.r
\name{message}
\alias{message}
\title{Message}
\usage{
message(...)
}
\arguments{
\item{req}{Request from api.ai}
}
\description{
This function parses messages to Rickett
}
\examples{
message()
}
\keyword{chat}
\keyword{finance}

