% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickett.r
\name{processMessage}
\alias{processMessage}
\title{Process Message}
\usage{
processMessage(req)
}
\arguments{
\item{req}{Request from api.ai}
}
\description{
Helper function to point request to correct parsing function
}

